requirements.txt
langchain
langchain-community
langchain-openai
openai
python-dotenv
pypdf